replace enum with frontend constants
