+++ src/constants/role.ts
@@ -1,0 +1,5 @@
+export enum Role {
+  USER = "USER",
+  ADMIN = "ADMIN",
+  SUPER_ADMIN = "SUPER_ADMIN",
+}
